requirements.txt
whisperx
pyannote.audio
PyQt5
torch
pyyaml